Only start DevTools server once
