+++ flutter-idea/src/io/flutter/devtools/RemainingDevToolsViewFactory.java
@@ -18,6 +18,7 @@ import io.flutter.sdk.FlutterSdkVersion;
 import io.flutter.utils.AsyncUtils;
 import io.flutter.utils.OpenApiUtils;
 import io.flutter.view.FlutterViewMessages;
+import io.flutter.view.ViewUtils;
 import kotlin.coroutines.Continuation;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -27,6 +28,9 @@ import java.util.Optional;
 
 public class RemainingDevToolsViewFactory implements ToolWindowFactory {
   @NotNull private static String TOOL_WINDOW_ID = "Flutter DevTools";
+
+  @NotNull
+  private final ViewUtils viewUtils = new ViewUtils();
 
   public static void init(Project project) {
     project.getMessageBus().connect().subscribe(
@@ -41,7 +45,6 @@ public class RemainingDevToolsViewFactory implements ToolWindowFactory {
     service.updateVmServiceUri(vmServiceUri);
   }
 
-
   @Override
   public void createToolWindowContent(@NotNull Project project, @NotNull ToolWindow window) {
     final ContentManager contentManager = window.getContentManager();
@@ -52,16 +55,8 @@ public class RemainingDevToolsViewFactory implements ToolWindowFactory {
     AsyncUtils.whenCompleteUiThread(
       DevToolsService.getInstance(project).getDevToolsInstance(),
       (instance, error) -> {
-        // Skip displaying if the project has been closed.
-        if (!project.isOpen()) {
-          return;
-        }
-
-        if (error != null) {
-          return;
-        }
-
-        if (instance == null) {
+        final boolean inValidState = viewUtils.checkDevToolsPanelInValidState(window, project, instance, error);
+        if (!inValidState) {
           return;
         }
