+++ flutter-idea/src/io/flutter/view/ViewUtils.java
@@ -5,6 +5,7 @@
  */
 package io.flutter.view;
 
+import com.intellij.openapi.project.Project;
 import com.intellij.openapi.ui.VerticalFlowLayout;
 import com.intellij.openapi.wm.ToolWindow;
 import com.intellij.ui.components.JBLabel;
@@ -13,6 +14,7 @@ import com.intellij.ui.content.Content;
 import com.intellij.ui.content.ContentManager;
 import com.intellij.util.ui.JBUI;
 import com.intellij.util.ui.UIUtil;
+import io.flutter.run.daemon.DevToolsInstance;
 import io.flutter.utils.LabelInput;
 import io.flutter.utils.OpenApiUtils;
 
@@ -22,7 +24,7 @@ import java.util.List;
 
 public class ViewUtils {
   public void presentLabel(ToolWindow toolWindow, String text) {
-    final JBLabel label = new JBLabel(text, SwingConstants.CENTER);
+    final JBLabel label = new JBLabel("<html>" + text + "</html>", SwingConstants.CENTER);
     label.setForeground(UIUtil.getLabelDisabledForeground());
     replacePanelLabel(toolWindow, label);
   }
@@ -49,6 +51,26 @@ public class ViewUtils {
     final JPanel center = new JPanel(new VerticalFlowLayout(VerticalFlowLayout.CENTER));
     center.add(panel);
     replacePanelLabel(toolWindow, center);
+  }
+
+  public boolean checkDevToolsPanelInValidState(ToolWindow toolWindow, Project project, DevToolsInstance instance, Throwable error) {
+    if (!project.isOpen()) {
+      presentLabel(toolWindow, "<h1>Project is not open.</h1>");
+      return false;
+    }
+
+    final String restartDevToolsMessage = "</br></br><h2>Try switching to another Flutter panel and back again to re-start the server.</h2>";
+    if (error != null) {
+      presentLabel(toolWindow, "<h1>Flutter DevTools start-up failed.</h1>" + restartDevToolsMessage);
+      return false;
+    }
+
+    if (instance == null) {
+      presentLabel(toolWindow, "<h1>Flutter DevTools does not exist.</h1>" + restartDevToolsMessage);
+      return false;
+    }
+
+    return true;
   }
 
   private void replacePanelLabel(ToolWindow toolWindow, JComponent label) {
